+++ src/Dispatcher.Web/HistoryContracts.cs
@@ -3,7 +3,7 @@ namespace Dispatcher.Web;
 public sealed record HistoryRecordPayload(
     ulong Position,
     string Kind,
-    long? Value,
+    decimal? Value,
     string? Unit,
     string? Quality,
     string? Freshness,
@@ -24,9 +24,9 @@ public sealed record HistoryBucketPayload(
     DateTimeOffset FromInclusive,
     DateTimeOffset ToExclusive,
     long Count,
-    double Average,
-    long Minimum,
-    long Maximum,
+    decimal Average,
+    decimal Minimum,
+    decimal Maximum,
     string Quality,
     string Freshness,
     bool HasGap);
