+++ tests/Dispatcher.IntegrationTests/RuntimeAlarmDeliveryProcessorTests.cs
@@ -394,7 +394,7 @@ public sealed class RuntimeAlarmDeliveryProcessorTests
                         SourceId,
                         PointId,
                         new OwnerPosition<SourceObservation>(1),
-                        TypedValue.From(value),
+                        TypedValue.From((decimal)value),
                         Unit.FromSymbol("kW"),
                         DataQuality.Good,
                         Freshness.Fresh,
